+++ backend/internal/library/producer/interface.go
@@ -1,0 +1,7 @@
+package producer
+
+import "context"
+
+type Producer interface {
+	Send(ctx context.Context, body string) error
+}
