+++ backend/internal/library/producer/send.go
@@ -1,4 +1,4 @@
-package sqs
+package producer
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/sqs"
 )
 
-func (p *publisher) SendMessage(ctx context.Context, body string) error {
+func (p *producer) Send(ctx context.Context, body string) error {
 	_, err := p.client.SendMessage(ctx, &sqs.SendMessageInput{
 		QueueUrl:    aws.String(p.queueURL),
 		MessageBody: aws.String(body),
