--- backend/internal/library/sqs/interface.go
@@ -1,7 +1,0 @@
-package sqs
-
-import "context"
-
-type Publisher interface {
-	SendMessage(ctx context.Context, body string) error
-}
