--- backend/internal/library/sqs/send_message_mock.go
@@ -1,13 +1,0 @@
-package sqs
-
-import (
-	"context"
-	"log"
-)
-
-type mockPublisher struct{}
-
-func (m *mockPublisher) SendMessage(ctx context.Context, body string) error {
-	log.Printf("[MOCK] SQS SendMessage: %s\n", body)
-	return nil
-}
